+++ crates/swc/tests/errors/lints/symbol-description/default/input.js
@@ -1,0 +1,4 @@
+Symbol()
+Symbol({})
+foo(() => Symbol())
+Symbol('name')