+++ crates/swc_ecma_minifier/tests/terser/compress/drop-unused/issue_2226_1/output.js
@@ -6,7 +6,6 @@ function f2(a) {
     b;
 }
 function f3(a) {
-    0;
 }
 function f4() {
     var a = b;